+++ pyVHDLModel/DesignUnit.py
@@ -39,12 +39,12 @@ from typing import List, Dict, Union, Iterable
 from pyTooling.Decorators import export
 from pyTooling.Graph import Vertex
 
-from pyVHDLModel import ModelEntity, NamedEntityMixin, DocumentedEntityMixin
-from pyVHDLModel.Symbol import NewSymbol, PackageSymbol, EntitySymbol
-from pyVHDLModel.Interface import GenericInterfaceItem, PortInterfaceItem
+from pyVHDLModel.Base       import ModelEntity, NamedEntityMixin, DocumentedEntityMixin
+from pyVHDLModel.Symbol     import NewSymbol, PackageSymbol, EntitySymbol
+from pyVHDLModel.Interface  import GenericInterfaceItem, PortInterfaceItem
 from pyVHDLModel.Subprogram import Procedure, Function
-from pyVHDLModel.Object import Constant, Variable, Signal
-from pyVHDLModel.Type import Type, Subtype
+from pyVHDLModel.Object     import Constant, Variable, Signal
+from pyVHDLModel.Type       import Type, Subtype
 from pyVHDLModel.Concurrent import ConcurrentStatement, ConcurrentStatements, ConcurrentDeclarations
