+++ pyVHDLModel/Interface.py
@@ -38,7 +38,8 @@ from typing import Iterable
 
 from pyTooling.Decorators import export
 
-from pyVHDLModel import DocumentedEntityMixin, Mode, SubtypeOrSymbol, ExpressionUnion
+from pyVHDLModel.Symbol import NewSymbol
+from pyVHDLModel.Base import DocumentedEntityMixin, ExpressionUnion, Mode
 from pyVHDLModel.Object import Constant, Signal, Variable, File
 from pyVHDLModel.Subprogram import Procedure, Function
 from pyVHDLModel.Type import Type
@@ -87,7 +88,7 @@ class ParameterInterfaceItem(InterfaceItem):
 
 @export
 class GenericConstantInterfaceItem(Constant, GenericInterfaceItem, InterfaceItemWithMode):
-	def __init__(self, identifiers: Iterable[str], mode: Mode, subtype: SubtypeOrSymbol, defaultExpression: ExpressionUnion = None, documentation: str = None):
+	def __init__(self, identifiers: Iterable[str], mode: Mode, subtype: NewSymbol, defaultExpression: ExpressionUnion = None, documentation: str = None):
 		super().__init__(identifiers, subtype, defaultExpression, documentation)
 		GenericInterfaceItem.__init__(self)
 		InterfaceItemWithMode.__init__(self, mode)
@@ -128,14 +129,14 @@ class GenericPackageInterfaceItem(GenericInterfaceItem):
 
 @export
 class PortSignalInterfaceItem(Signal, PortInterfaceItem):
-	def __init__(self, identifiers: Iterable[str], mode: Mode, subtype: SubtypeOrSymbol, defaultExpression: ExpressionUnion = None, documentation: str = None):
+	def __init__(self, identifiers: Iterable[str], mode: Mode, subtype: NewSymbol, defaultExpression: ExpressionUnion = None, documentation: str = None):
 		super().__init__(identifiers, subtype, defaultExpression, documentation)
 		PortInterfaceItem.__init__(self, mode)
 
 
 @export
 class ParameterConstantInterfaceItem(Constant, ParameterInterfaceItem, InterfaceItemWithMode):
-	def __init__(self, identifiers: Iterable[str], mode: Mode, subtype: SubtypeOrSymbol, defaultExpression: ExpressionUnion = None, documentation: str = None):
+	def __init__(self, identifiers: Iterable[str], mode: Mode, subtype: NewSymbol, defaultExpression: ExpressionUnion = None, documentation: str = None):
 		super().__init__(identifiers, subtype, defaultExpression, documentation)
 		ParameterInterfaceItem.__init__(self)
 		InterfaceItemWithMode.__init__(self, mode)
@@ -143,7 +144,7 @@ class ParameterConstantInterfaceItem(Constant, ParameterInterfaceItem, Interface
 
 @export
 class ParameterVariableInterfaceItem(Variable, ParameterInterfaceItem, InterfaceItemWithMode):
-	def __init__(self, identifiers: Iterable[str], mode: Mode, subtype: SubtypeOrSymbol, defaultExpression: ExpressionUnion = None, documentation: str = None):
+	def __init__(self, identifiers: Iterable[str], mode: Mode, subtype: NewSymbol, defaultExpression: ExpressionUnion = None, documentation: str = None):
 		super().__init__(identifiers, subtype, defaultExpression, documentation)
 		ParameterInterfaceItem.__init__(self)
 		InterfaceItemWithMode.__init__(self, mode)
@@ -151,7 +152,7 @@ class ParameterVariableInterfaceItem(Variable, ParameterInterfaceItem, Interface
 
 @export
 class ParameterSignalInterfaceItem(Signal, ParameterInterfaceItem, InterfaceItemWithMode):
-	def __init__(self, identifiers: Iterable[str], mode: Mode, subtype: SubtypeOrSymbol, defaultExpression: ExpressionUnion = None, documentation: str = None):
+	def __init__(self, identifiers: Iterable[str], mode: Mode, subtype: NewSymbol, defaultExpression: ExpressionUnion = None, documentation: str = None):
 		super().__init__(identifiers, subtype, defaultExpression, documentation)
 		ParameterInterfaceItem.__init__(self)
 		InterfaceItemWithMode.__init__(self, mode)
@@ -159,6 +160,6 @@ class ParameterSignalInterfaceItem(Signal, ParameterInterfaceItem, InterfaceItem
 
 @export
 class ParameterFileInterfaceItem(File, ParameterInterfaceItem):
-	def __init__(self, identifiers: Iterable[str], subtype: SubtypeOrSymbol, documentation: str = None):
+	def __init__(self, identifiers: Iterable[str], subtype: NewSymbol, documentation: str = None):
 		super().__init__(identifiers, subtype, documentation)
 		ParameterInterfaceItem.__init__(self)
