+++ pyVHDLModel/Object.py
@@ -38,14 +38,15 @@ from typing import Iterable, Optional as Nullable
 
 from pyTooling.Decorators import export
 
-from pyVHDLModel import ModelEntity, MultipleNamedEntityMixin, DocumentedEntityMixin, SubtypeOrSymbol, ExpressionUnion
+from pyVHDLModel.Symbol import NewSymbol
+from pyVHDLModel.Base import ModelEntity, MultipleNamedEntityMixin, DocumentedEntityMixin, ExpressionUnion
 
 
 @export
 class Obj(ModelEntity, MultipleNamedEntityMixin, DocumentedEntityMixin):
-	_subtype: SubtypeOrSymbol
+	_subtype: NewSymbol
 
-	def __init__(self, identifiers: Iterable[str], subtype: SubtypeOrSymbol, documentation: str = None):
+	def __init__(self, identifiers: Iterable[str], subtype: NewSymbol, documentation: str = None):
 		super().__init__()
 		MultipleNamedEntityMixin.__init__(self, identifiers)
 		DocumentedEntityMixin.__init__(self, documentation)
@@ -54,7 +55,7 @@ class Obj(ModelEntity, MultipleNamedEntityMixin, DocumentedEntityMixin):
 		subtype._parent = self
 
 	@property
-	def Subtype(self) -> SubtypeOrSymbol:
+	def Subtype(self) -> NewSymbol:
 		return self._subtype
 
 
@@ -81,7 +82,7 @@ class WithDefaultExpressionMixin:
 
 @export
 class Constant(BaseConstant, WithDefaultExpressionMixin):
-	def __init__(self, identifiers: Iterable[str], subtype: SubtypeOrSymbol, defaultExpression: ExpressionUnion = None, documentation: str = None):
+	def __init__(self, identifiers: Iterable[str], subtype: NewSymbol, defaultExpression: ExpressionUnion = None, documentation: str = None):
 		super().__init__(identifiers, subtype, documentation)
 		WithDefaultExpressionMixin.__init__(self, defaultExpression)
 
@@ -90,7 +91,7 @@ class Constant(BaseConstant, WithDefaultExpressionMixin):
 class DeferredConstant(BaseConstant):
 	_constantReference: Constant
 
-	def __init__(self, identifiers: Iterable[str], subtype: SubtypeOrSymbol, documentation: str = None):
+	def __init__(self, identifiers: Iterable[str], subtype: NewSymbol, documentation: str = None):
 		super().__init__(identifiers, subtype, documentation)
 
 	@property
@@ -100,7 +101,7 @@ class DeferredConstant(BaseConstant):
 
 @export
 class Variable(Obj, WithDefaultExpressionMixin):
-	def __init__(self, identifiers: Iterable[str], subtype: SubtypeOrSymbol, defaultExpression: ExpressionUnion = None, documentation: str = None):
+	def __init__(self, identifiers: Iterable[str], subtype: NewSymbol, defaultExpression: ExpressionUnion = None, documentation: str = None):
 		super().__init__(identifiers, subtype, documentation)
 		WithDefaultExpressionMixin.__init__(self, defaultExpression)
 
@@ -112,7 +113,7 @@ class SharedVariable(Obj):
 
 @export
 class Signal(Obj, WithDefaultExpressionMixin):
-	def __init__(self, identifiers: Iterable[str], subtype: SubtypeOrSymbol, defaultExpression: ExpressionUnion = None, documentation: str = None):
+	def __init__(self, identifiers: Iterable[str], subtype: NewSymbol, defaultExpression: ExpressionUnion = None, documentation: str = None):
 		super().__init__(identifiers, subtype, documentation)
 		WithDefaultExpressionMixin.__init__(self, defaultExpression)
