+++ pyVHDLModel/Type.py
@@ -38,7 +38,7 @@ from typing                 import Union, List, Iterator, Iterable, Tuple
 
 from pyTooling.Decorators   import export
 
-from pyVHDLModel            import ModelEntity, NamedEntityMixin, MultipleNamedEntityMixin, DocumentedEntityMixin, ExpressionUnion
+from pyVHDLModel.Base import ModelEntity, NamedEntityMixin, MultipleNamedEntityMixin, DocumentedEntityMixin, ExpressionUnion
 from pyVHDLModel.Symbol import NewSymbol
 from pyVHDLModel.Name       import Name
 from pyVHDLModel.Common     import Range
